requirements.txt
wxPython
PyOpenGL
opencv-python
keyboard
pyrr
serial
